Move execute_on function to cli-xtask
