+++ xtask/src/udeps.rs
@@ -1,4 +1,4 @@
-use cli_xtask::workspace;
+use cli_xtask::{process, workspace};
 
 #[derive(Debug, clap::Parser)]
 pub(crate) struct Args {
@@ -15,7 +15,7 @@ impl Args {
             for package in metadata.workspace_packages() {
                 for feature_args in crate::feature_combinations(package) {
                     // cargo +nightly udeps fails on windows, so use rustup instead
-                    crate::execute_on(
+                    process::execute_on(
                         metadata,
                         "rustup",
                         [
